Add wrapper script for git init.
Sets (globally or locally):
1. author
2. email address
3. time zone
4. aliases
   - alt-commit
   - alt-revert

The time zone is set using timedatectl if the value is not provided on
the command line.

The aliases are wrappers around 'commit' and 'revert' that alter the
GIT_COMMITTER_DATE and GIT_AUTHOR_DATE environmental variables.

The aliases can be used for two types of 'schedules':
 * workday: set start and end time for workday, and the alias will
   compress the current time (24-hour window) to the corresponding time
   for the workday window (start - end)
 * random: if concerned about git commits leaking metadata about your
   work schedule, you can add an offset (e.g. 3 days, 2 weeks) to
   'smear' the commit time along the offset, centered on the present
   time. The more frequent and closely spaced your commits, the more
   broadly you should set the interval.