[count-lines] add wrapper function to strip deleted files
`git ls-files` will return names for files that are deleted on the
filesystem but not deleted in the git index. `find` will spit out an
error but not fail, but prevent non-existing files from being sent in
the first place, add a wrapper function `strip-missing` that removes all
the deleted files (`ls-files -d`) from the `ls-files -co` string.

Because the string is nul-terminated, the pattern is "s/$file\x00//",
where x00 is the null byte (for sed). `perl -pe "s/${i}\000//g"` would
also work.