[count-commits] add special case when called inside a git repo
If called from within a subdirectory of a Git repository, there is no
.git/ directory to be found, causing count-commits to return empty.

If '.' is passed as the directory and `git log` returns successfully,
the directory is part of a Git repository. In this case, the script will
call get_count directly. Otherwise, call get_count via a find subshell
as before.

This fix works because  `git rev-list` can be run anywhere in a Git
repository.